requirements.txt
python-telegram-bot==20.7
yt-dlp==2024.4.9
flask==3.0.0
gunicorn==21.2.0
